@@ -25,10 +25,6 @@
 # and warranty status of this software.
 
 '''Script-related classes and functions.'''
-
-
-from collections import namedtuple
-
 from electrumx.lib.enum import Enumeration
 from electrumx.lib.util import unpack_le_uint16_from, unpack_le_uint32_from, \
     pack_le_uint16, pack_le_uint32
@@ -100,62 +96,30 @@ def _match_ops(ops, pattern):
     return True
 
 
-class ScriptPubKey(object):
+class ScriptPubKey:
     '''A class for handling a tx output script that gives conditions
     necessary for spending.
     '''
 
-    TO_ADDRESS_OPS = [OpCodes.OP_DUP, OpCodes.OP_HASH160, -1,
-                      OpCodes.OP_EQUALVERIFY, OpCodes.OP_CHECKSIG]
-    TO_P2SH_OPS = [OpCodes.OP_HASH160, -1, OpCodes.OP_EQUAL]
-    TO_PUBKEY_OPS = [-1, OpCodes.OP_CHECKSIG]
-
-    PayToHandlers = namedtuple('PayToHandlers', 'address script_hash pubkey '
-                               'unspendable strange')
-
-    @classmethod
-    def pay_to(cls, handlers, script):
-        '''Parse a script, invoke the appropriate handler and
-        return the result.
-
-        One of the following handlers is invoked:
-           handlers.address(hash160)
-           handlers.script_hash(hash160)
-           handlers.pubkey(pubkey)
-           handlers.unspendable()
-           handlers.strange(script)
-        '''
-        try:
-            ops = Script.get_ops(script)
-        except ScriptError:
-            return handlers.unspendable()
-
-        match = _match_ops
-
-        if match(ops, cls.TO_ADDRESS_OPS):
-            return handlers.address(ops[2][-1])
-        if match(ops, cls.TO_P2SH_OPS):
-            return handlers.script_hash(ops[1][-1])
-        if match(ops, cls.TO_PUBKEY_OPS):
-            return handlers.pubkey(ops[0][-1])
-        if ops and ops[0] == OpCodes.OP_RETURN:
-            return handlers.unspendable()
-        return handlers.strange(script)
+    TO_ADDRESS_OPS = (OpCodes.OP_DUP, OpCodes.OP_HASH160, -1,
+                      OpCodes.OP_EQUALVERIFY, OpCodes.OP_CHECKSIG)
+    TO_P2SH_OPS = (OpCodes.OP_HASH160, -1, OpCodes.OP_EQUAL)
+    TO_PUBKEY_OPS = (-1, OpCodes.OP_CHECKSIG)
 
     @classmethod
     def P2SH_script(cls, hash160):
-        return (bytes([OpCodes.OP_HASH160])
+        return (bytes((OpCodes.OP_HASH160,))
                 + Script.push_data(hash160)
-                + bytes([OpCodes.OP_EQUAL]))
+                + bytes((OpCodes.OP_EQUAL,)))
 
     @classmethod
     def P2PKH_script(cls, hash160):
-        return (bytes([OpCodes.OP_DUP, OpCodes.OP_HASH160])
+        return (bytes((OpCodes.OP_DUP, OpCodes.OP_HASH160))
                 + Script.push_data(hash160)
-                + bytes([OpCodes.OP_EQUALVERIFY, OpCodes.OP_CHECKSIG]))
+                + bytes((OpCodes.OP_EQUALVERIFY, OpCodes.OP_CHECKSIG)))
 
 
-class Script(object):
+class Script:
 
     @classmethod
     def get_ops(cls, script):
@@ -201,21 +165,21 @@ class Script(object):
 
         n = len(data)
         if n < OpCodes.OP_PUSHDATA1:
-            return bytes([n]) + data
+            return bytes((n,)) + data
         if n < 256:
-            return bytes([OpCodes.OP_PUSHDATA1, n]) + data
+            return bytes((OpCodes.OP_PUSHDATA1, n)) + data
         if n < 65536:
-            return bytes([OpCodes.OP_PUSHDATA2]) + pack_le_uint16(n) + data
-        return bytes([OpCodes.OP_PUSHDATA4]) + pack_le_uint32(n) + data
+            return bytes((OpCodes.OP_PUSHDATA2,)) + pack_le_uint16(n) + data
+        return bytes((OpCodes.OP_PUSHDATA4,)) + pack_le_uint32(n) + data
 
     @classmethod
     def opcode_name(cls, opcode):
         if OpCodes.OP_0 < opcode < OpCodes.OP_PUSHDATA1:
-            return 'OP_{:d}'.format(opcode)
+            return f'OP_{opcode:d}'
         try:
             return OpCodes.whatis(opcode)
         except KeyError:
-            return 'OP_UNKNOWN:{:d}'.format(opcode)
+            return f'OP_UNKNOWN:{opcode:d}'
 
     @classmethod
     def dump(cls, script):
@@ -225,5 +189,4 @@ class Script(object):
             if data is None:
                 print(name)
             else:
-                print('{} {} ({:d} bytes)'
-                      .format(name, data.hex(), len(data)))
+                print(f'{name} {data.hex()} ({len(data):d} bytes)')
