@@ -27,15 +27,29 @@
 '''Miscellaneous utility classes and functions.'''
 
 
-import array
+from array import array
+import asyncio
 import inspect
 from ipaddress import ip_address
 import logging
-import re
 import sys
 from collections.abc import Container, Mapping
 from struct import Struct
 
+import aiorpcx
+
+
+# Use system-compiled JSON lib if available, fallback to stdlib
+try:
+    import rapidjson as json
+except ImportError:
+    try:
+        import ujson as json
+    except ImportError:
+        import json
+
+json_deserialize = json.loads
+json_serialize = json.dumps
 
 # Logging utilities
 
@@ -71,8 +85,7 @@ def class_logger(path, classname):
 # Method decorator.  To be used for calculations that will always
 # deliver the same result.  The method cannot take any arguments
 # and should be accessed as an attribute.
-class cachedproperty(object):
-
+class cachedproperty:
     def __init__(self, f):
         self.f = f
 
@@ -95,7 +108,7 @@ def formatted_time(t, sep=' '):
             parts.append(fmt.format(val))
         t %= n
     if len(parts) < 3:
-        parts.append('{:02d}s'.format(t))
+        parts.append(f'{t:02d}s')
     return sep.join(parts)
 
 
@@ -122,7 +135,7 @@ def deep_getsizeof(obj):
         r = sys.getsizeof(o)
         ids.add(id(o))
 
-        if isinstance(o, (str, bytes, bytearray, array.array)):
+        if isinstance(o, (str, bytes, bytearray, array)):
             return r
 
         if isinstance(o, Mapping):
@@ -153,9 +166,9 @@ def chunks(items, size):
 
 
 def resolve_limit(limit):
-    if limit is None:
+    if limit is None or limit < 0:
         return -1
-    assert isinstance(limit, int) and limit >= 0
+    assert isinstance(limit, int)
     return limit
 
 
@@ -173,18 +186,17 @@ def increment_byte_string(bs):
     '''Return the lexicographically next byte string of the same length.
 
     Return None if there is none (when the input is all 0xff bytes).'''
-    for n in range(1, len(bs) + 1):
-        if bs[-n] != 0xff:
-            return bs[:-n] + bytes([bs[-n] + 1]) + bytes(n - 1)
-    return None
+    try:
+        return (int.from_bytes(bs, 'big') + 1).to_bytes(len(bs), 'big')
+    except OverflowError:
+        return None
 
 
-class LogicalFile(object):
+class LogicalFile:
     '''A logical binary file split across several separate files on disk.'''
 
     def __init__(self, prefix, digits, file_size):
-        digit_fmt = '{' + ':0{:d}d'.format(digits) + '}'
-        self.filename_fmt = prefix + digit_fmt
+        self.filename_fmt = f'{prefix}{{:0{digits:d}d}}'
         self.file_size = file_size
 
     def read(self, start, size=-1):
@@ -245,7 +257,6 @@ def open_truncate(filename):
 
 def address_string(address):
     '''Return an address as a correctly formatted string.'''
-    fmt = '{}:{:d}'
     host, port = address
     try:
         host = ip_address(host)
@@ -253,8 +264,8 @@ def address_string(address):
         pass
     else:
         if host.version == 6:
-            fmt = '[{}]:{:d}'
-    return fmt.format(host, port)
+            return f'[{host}]:{port:d}'
+    return f'{host}:{port:d}'
 
 
 def protocol_tuple(s):
@@ -346,3 +357,57 @@ def pack_varint(n):
 
 def pack_varbytes(data):
     return pack_varint(len(data)) + data
+
+
+class OldTaskGroup(aiorpcx.TaskGroup):
+    """Automatically raises exceptions on join; as in aiorpcx prior to version 0.20"""
+    async def join(self):
+        if self._wait is all:
+            exc = False
+            try:
+                async for task in self:
+                    if not task.cancelled():
+                        task.result()
+            except BaseException:  # including asyncio.CancelledError
+                exc = True
+                raise
+            finally:
+                if exc:
+                    await self.cancel_remaining()
+                await super().join()
+        else:
+            await super().join()
+            if self.completed:
+                self.completed.result()
+
+
+# We monkey-patch aiorpcx TimeoutAfter (used by timeout_after and ignore_after API),
+# to fix a timing issue present in asyncio as a whole re timing out tasks.
+# To see the issue we are trying to fix, consider example:
+#     async def outer_task():
+#         async with timeout_after(0.1):
+#             await inner_task()
+# When the 0.1 sec timeout expires, inner_task will get cancelled by timeout_after (=internal cancellation).
+# If around the same time (in terms of event loop iterations) another coroutine
+# cancels outer_task (=external cancellation), there will be a race.
+# Both cancellations work by propagating a CancelledError out to timeout_after, which then
+# needs to decide (in TimeoutAfter.__aexit__) whether it's due to an internal or external cancellation.
+# AFAICT asyncio provides no reliable way of distinguishing between the two.
+# This patch tries to always give priority to external cancellations.
+# see https://github.com/kyuupichan/aiorpcX/issues/44
+# see https://github.com/aio-libs/async-timeout/issues/229
+# see https://bugs.python.org/issue42130 and https://bugs.python.org/issue45098
+def _aiorpcx_monkeypatched_set_new_deadline(task, deadline):
+    def timeout_task():
+        task._orig_cancel()
+        task._timed_out = None if getattr(task, "_externally_cancelled", False) else deadline
+    def mycancel(*args, **kwargs):
+        task._orig_cancel(*args, **kwargs)
+        task._externally_cancelled = True
+        task._timed_out = None
+    if not hasattr(task, "_orig_cancel"):
+        task._orig_cancel = task.cancel
+        task.cancel = mycancel
+    task._deadline_handle = task._loop.call_at(deadline, timeout_task)
+
+aiorpcx.curio._set_new_deadline = _aiorpcx_monkeypatched_set_new_deadline
