@@ -10,18 +10,22 @@ daemon.'''
 
 import asyncio
 import itertools
-import json
 import time
 from calendar import timegm
 from struct import pack
+from typing import TYPE_CHECKING, Type
 
 import aiohttp
 from aiorpcx import JSONRPC
 
-from electrumx.lib.util import hex_to_bytes, class_logger,\
-    unpack_le_uint16_from, pack_varint
-from electrumx.lib.hash import hex_str_to_hash, hash_to_hex_str
+from electrumx.lib.hash import hash_to_hex_str, hex_str_to_hash
 from electrumx.lib.tx import DeserializerDecred
+from electrumx.lib.util import (class_logger, hex_to_bytes, json_deserialize,
+                                json_serialize, pack_varint,
+                                unpack_le_uint16_from)
+
+if TYPE_CHECKING:
+    from electrumx.lib.coins import Coin
 
 
 class DaemonError(Exception):
@@ -37,13 +41,21 @@ class ServiceRefusedError(Exception):
     some reason.'''
 
 
-class Daemon(object):
+class Daemon:
     '''Handles connections to a daemon at the given URL.'''
 
     WARMING_UP = -28
     id_counter = itertools.count()
 
-    def __init__(self, coin, url, *, max_workqueue=10, init_retry=0.25, max_retry=4.0):
+    def __init__(
+            self,
+            coin: Type['Coin'],
+            url,
+            *,
+            max_workqueue=10,
+            init_retry=0.25,
+            max_retry=4.0,
+    ):
         self.coin = coin
         self.logger = class_logger(__name__, self.__class__.__name__)
         self.url_index = None
@@ -57,6 +69,9 @@ class Daemon(object):
         self._height = None
         self.available_rpcs = {}
         self.session = None
+
+        self._networkinfo_cache = (None, 0)
+        self._networkinfo_lock = asyncio.Lock()
 
     async def __aenter__(self):
         self.session = aiohttp.ClientSession(connector=self.connector())
@@ -105,7 +120,7 @@ class Daemon(object):
             async with self.session.post(self.current_url(), data=data) as resp:
                 kind = resp.headers.get('Content-Type', None)
                 if kind == 'application/json':
-                    return await resp.json()
+                    return await resp.json(loads=json_deserialize)
                 text = await resp.text()
                 text = text.strip() or resp.reason
                 raise ServiceRefusedError(text)
@@ -118,7 +133,7 @@ class Daemon(object):
         '''
         def log_error(error):
             nonlocal last_error_log, retry
-            now = time.time()
+            now = time.monotonic()
             if now - last_error_log > 60:
                 last_error_log = now
                 self.logger.error(f'{error}.  Retrying occasionally...')
@@ -127,7 +142,7 @@ class Daemon(object):
 
         on_good_message = None
         last_error_log = 0
-        data = json.dumps(payload)
+        data = json_serialize(payload)
         retry = self.init_retry
         while True:
             try:
@@ -232,12 +247,15 @@ class Daemon(object):
         '''Update our record of the daemon's mempool hashes.'''
         return await self._send_single('getrawmempool')
 
-    async def estimatefee(self, block_count):
+    async def estimatefee(self, block_count, estimate_mode=None):
         '''Return the fee estimate for the block count.  Units are whole
         currency units per KB, e.g. 0.00000995, or -1 if no estimate
         is available.
         '''
-        args = (block_count, )
+        if estimate_mode:
+            args = (block_count, estimate_mode)
+        else:
+            args = (block_count, )
         if await self._is_rpc_available('estimatesmartfee'):
             estimate = await self._send_single('estimatesmartfee', args)
             return estimate.get('feerate', -1)
@@ -245,7 +263,13 @@ class Daemon(object):
 
     async def getnetworkinfo(self):
         '''Return the result of the 'getnetworkinfo' RPC call.'''
-        return await self._send_single('getnetworkinfo')
+        async with self._networkinfo_lock:
+            cache_val, cache_time = self._networkinfo_cache
+            if time.time() - cache_time < 60:  # seconds
+                return cache_val
+            val = await self._send_single('getnetworkinfo')
+            self._networkinfo_cache = (val, time.time())
+            return val
 
     async def relayfee(self):
         '''The minimum fee a low-priority tx must pay in order to be accepted
