@@ -8,35 +8,55 @@
 
 '''History by script hash (address).'''
 
-import array
 import ast
 import bisect
 import time
+from array import array
 from collections import defaultdict
-from functools import partial
+from typing import TYPE_CHECKING, Type, Optional
 
 import electrumx.lib.util as util
-from electrumx.lib.util import pack_be_uint16, unpack_be_uint16_from
-from electrumx.lib.hash import hash_to_hex_str, HASHX_LEN
+from electrumx.lib.hash import HASHX_LEN, hash_to_hex_str
+from electrumx.lib.util import (pack_be_uint16, pack_le_uint64,
+                                unpack_be_uint16_from, unpack_le_uint64)
+
+if TYPE_CHECKING:
+    from electrumx.server.storage import Storage
 
 
-class History(object):
+TXNUM_LEN = 5
+FLUSHID_LEN = 2
 
-    DB_VERSIONS = [0]
+
+class History:
+
+    DB_VERSIONS = (0, 1)
+
+    db: Optional['Storage']
 
     def __init__(self):
         self.logger = util.class_logger(__name__, self.__class__.__name__)
         # For history compaction
         self.max_hist_row_entries = 12500
-        self.unflushed = defaultdict(partial(array.array, 'I'))
+        self.unflushed = defaultdict(bytearray)
         self.unflushed_count = 0
         self.flush_count = 0
         self.comp_flush_count = -1
         self.comp_cursor = -1
         self.db_version = max(self.DB_VERSIONS)
+        self.upgrade_cursor = -1
+
+        # Key: address_hashX + flush_id
+        # Value: sorted "list" of tx_nums in history of hashX
         self.db = None
 
-    def open_db(self, db_class, for_sync, utxo_flush_count, compacting):
+    def open_db(
+            self,
+            db_class: Type['Storage'],
+            for_sync: bool,
+            utxo_flush_count: int,
+            compacting: bool,
+    ):
         self.db = db_class('hist', for_sync)
         self.read_state()
         self.clear_excess(utxo_flush_count)
@@ -61,17 +81,22 @@ class History(object):
             self.comp_flush_count = state.get('comp_flush_count', -1)
             self.comp_cursor = state.get('comp_cursor', -1)
             self.db_version = state.get('db_version', 0)
+            self.upgrade_cursor = state.get('upgrade_cursor', -1)
         else:
             self.flush_count = 0
             self.comp_flush_count = -1
             self.comp_cursor = -1
             self.db_version = max(self.DB_VERSIONS)
+            self.upgrade_cursor = -1
 
-        self.logger.info(f'history DB version: {self.db_version}')
         if self.db_version not in self.DB_VERSIONS:
-            msg = f'this software only handles DB versions {self.DB_VERSIONS}'
+            msg = (f'your history DB version is {self.db_version} but '
+                   f'this software only handles DB versions {self.DB_VERSIONS}')
             self.logger.error(msg)
             raise RuntimeError(msg)
+        if self.db_version != max(self.DB_VERSIONS):
+            self.upgrade_db()
+        self.logger.info(f'history DB version: {self.db_version}')
         self.logger.info(f'flush count: {self.flush_count:,d}')
 
     def clear_excess(self, utxo_flush_count):
@@ -85,7 +110,7 @@ class History(object):
 
         keys = []
         for key, _hist in self.db.iterator(prefix=b''):
-            flush_id, = unpack_be_uint16_from(key[-2:])
+            flush_id, = unpack_be_uint16_from(key[-FLUSHID_LEN:])
             if flush_id > utxo_flush_count:
                 keys.append(key)
 
@@ -106,6 +131,7 @@ class History(object):
             'comp_flush_count': self.comp_flush_count,
             'comp_cursor': self.comp_cursor,
             'db_version': self.db_version,
+            'upgrade_cursor': self.upgrade_cursor,
         }
         # History entries are not prefixed; the suffix \0\0 ensures we
         # look similar to other entries and aren't interfered with
@@ -115,20 +141,21 @@ class History(object):
         unflushed = self.unflushed
         count = 0
         for tx_num, hashXs in enumerate(hashXs_by_tx, start=first_tx_num):
+            tx_numb = pack_le_uint64(tx_num)[:TXNUM_LEN]
             hashXs = set(hashXs)
             for hashX in hashXs:
-                unflushed[hashX].append(tx_num)
+                unflushed[hashX] += tx_numb
             count += len(hashXs)
         self.unflushed_count += count
 
     def unflushed_memsize(self):
-        return len(self.unflushed) * 180 + self.unflushed_count * 4
+        return len(self.unflushed) * 180 + self.unflushed_count * TXNUM_LEN
 
     def assert_flushed(self):
         assert not self.unflushed
 
     def flush(self):
-        start_time = time.time()
+        start_time = time.monotonic()
         self.flush_count += 1
         flush_id = pack_be_uint16(self.flush_count)
         unflushed = self.unflushed
@@ -136,7 +163,7 @@ class History(object):
         with self.db.write_batch() as batch:
             for hashX in sorted(unflushed):
                 key = hashX + flush_id
-                batch.put(key, unflushed[hashX].tobytes())
+                batch.put(key, bytes(unflushed[hashX]))
             self.write_state(batch)
 
         count = len(unflushed)
@@ -144,7 +171,7 @@ class History(object):
         self.unflushed_count = 0
 
         if self.db.for_sync:
-            elapsed = time.time() - start_time
+            elapsed = time.monotonic() - start_time
             self.logger.info(f'flushed history in {elapsed:.1f}s '
                              f'for {count:,d} addrs')
 
@@ -153,19 +180,23 @@ class History(object):
         self.flush_count += 1
         nremoves = 0
         bisect_left = bisect.bisect_left
+        chunks = util.chunks
 
+        txnum_padding = bytes(8-TXNUM_LEN)
         with self.db.write_batch() as batch:
             for hashX in sorted(hashXs):
                 deletes = []
                 puts = {}
                 for key, hist in self.db.iterator(prefix=hashX, reverse=True):
-                    a = array.array('I')
-                    a.frombytes(hist)
+                    a = array(
+                        'Q',
+                        b''.join(item + txnum_padding for item in chunks(hist, TXNUM_LEN))
+                    )
                     # Remove all history entries >= tx_count
                     idx = bisect_left(a, tx_count)
                     nremoves += len(a) - idx
                     if idx > 0:
-                        puts[key] = a[:idx].tobytes()
+                        puts[key] = hist[:TXNUM_LEN * idx]
                         break
                     deletes.append(key)
 
@@ -183,12 +214,13 @@ class History(object):
         transactions.  By default yields at most 1000 entries.  Set
         limit to None to get them all.  '''
         limit = util.resolve_limit(limit)
+        chunks = util.chunks
+        txnum_padding = bytes(8-TXNUM_LEN)
         for _key, hist in self.db.iterator(prefix=hashX):
-            a = array.array('I')
-            a.frombytes(hist)
-            for tx_num in a:
+            for tx_numb in chunks(hist, TXNUM_LEN):
                 if limit == 0:
                     return
+                tx_num, = unpack_le_uint64(tx_numb + txnum_padding)
                 yield tx_num
                 limit -= 1
 
@@ -254,18 +286,18 @@ class History(object):
                        write_items, keys_to_delete):
         '''Compres history for a hashX.  hist_list is an ordered list of
         the histories to be compressed.'''
-        # History entries (tx numbers) are 4 bytes each.  Distribute
+        # History entries (tx numbers) are TXNUM_LEN bytes each.  Distribute
         # over rows of up to 50KB in size.  A fixed row size means
         # future compactions will not need to update the first N - 1
         # rows.
-        max_row_size = self.max_hist_row_entries * 4
+        max_row_size = self.max_hist_row_entries * TXNUM_LEN
         full_hist = b''.join(hist_list)
         nrows = (len(full_hist) + max_row_size - 1) // max_row_size
         if nrows > 4:
-            self.logger.info('hashX {} is large: {:,d} entries across '
-                             '{:,d} rows'
-                             .format(hash_to_hex_str(hashX),
-                                     len(full_hist) // 4, nrows))
+            self.logger.info(
+                f'hashX {hash_to_hex_str(hashX)} is large: '
+                f'{len(full_hist) // TXNUM_LEN:,d} entries across {nrows:,d} rows'
+            )
 
         # Find what history needs to be written, and what keys need to
         # be deleted.  Start by assuming all keys are to be deleted,
@@ -293,13 +325,13 @@ class History(object):
         hist_map = {}
         hist_list = []
 
-        key_len = HASHX_LEN + 2
+        key_len = HASHX_LEN + FLUSHID_LEN
         write_size = 0
         for key, hist in self.db.iterator(prefix=prefix):
             # Ignore non-history entries
             if len(key) != key_len:
                 continue
-            hashX = key[:-2]
+            hashX = key[:-FLUSHID_LEN]
             if hashX != prior_hashX and prior_hashX:
                 write_size += self._compact_hashX(prior_hashX, hist_map,
                                                   hist_list, write_items,
@@ -334,11 +366,12 @@ class History(object):
         max_rows = self.comp_flush_count + 1
         self._flush_compaction(cursor, write_items, keys_to_delete)
 
-        self.logger.info('history compaction: wrote {:,d} rows ({:.1f} MB), '
-                         'removed {:,d} rows, largest: {:,d}, {:.1f}% complete'
-                         .format(len(write_items), write_size / 1000000,
-                                 len(keys_to_delete), max_rows,
-                                 100 * cursor / 65536))
+        self.logger.info(
+            f'history compaction: wrote {len(write_items):,d} rows '
+            f'({write_size / 1000000:.1f} MB), removed '
+            f'{len(keys_to_delete):,d} rows, largest: {max_rows:,d}, '
+            f'{100 * cursor / 65536:.1f}% complete'
+        )
         return write_size
 
     def _cancel_compaction(self):
@@ -346,3 +379,46 @@ class History(object):
             self.logger.warning('cancelling in-progress history compaction')
             self.comp_flush_count = -1
             self.comp_cursor = -1
+
+    #
+    # DB upgrade
+    #
+
+    def upgrade_db(self):
+        self.logger.info(f'history DB version: {self.db_version}')
+        self.logger.info('Upgrading your history DB; this can take some time...')
+
+        def upgrade_cursor(cursor):
+            count = 0
+            prefix = pack_be_uint16(cursor)
+            key_len = HASHX_LEN + 2
+            chunks = util.chunks
+            with self.db.write_batch() as batch:
+                batch_put = batch.put
+                for key, hist in self.db.iterator(prefix=prefix):
+                    # Ignore non-history entries
+                    if len(key) != key_len:
+                        continue
+                    count += 1
+                    hist = b''.join(item + b'\0' for item in chunks(hist, 4))
+                    batch_put(key, hist)
+                self.upgrade_cursor = cursor
+                self.write_state(batch)
+            return count
+
+        last = time.monotonic()
+        count = 0
+
+        for cursor in range(self.upgrade_cursor + 1, 65536):
+            count += upgrade_cursor(cursor)
+            now = time.monotonic()
+            if now > last + 10:
+                last = now
+                self.logger.info(f'DB 3 of 3: {count:,d} entries updated, '
+                                 f'{cursor * 100 / 65536:.1f}% complete')
+
+        self.db_version = max(self.DB_VERSIONS)
+        self.upgrade_cursor = -1
+        with self.db.write_batch() as batch:
+            self.write_state(batch)
+        self.logger.info('DB 3 of 3 upgraded successfully')
