@@ -9,20 +9,21 @@
 
 import os
 from functools import partial
+from typing import Type
 
 import electrumx.lib.util as util
 
 
-def db_class(name):
+def db_class(name) -> Type['Storage']:
     '''Returns a DB engine class.'''
     for db_class in util.subclasses(Storage):
         if db_class.__name__.lower() == name.lower():
             db_class.import_module()
             return db_class
-    raise RuntimeError('unrecognised DB engine "{}"'.format(name))
+    raise RuntimeError(f'unrecognised DB engine "{name}"')
 
 
-class Storage(object):
+class Storage:
     '''Abstract base class of the DB backend abstraction.'''
 
     def __init__(self, name, for_sync):
@@ -123,7 +124,7 @@ class RocksDB(Storage):
         return RocksDBIterator(self.db, prefix, reverse)
 
 
-class RocksDBWriteBatch(object):
+class RocksDBWriteBatch:
     '''A write batch for RocksDB.'''
 
     def __init__(self, db):
@@ -138,7 +139,7 @@ class RocksDBWriteBatch(object):
             self.db.write(self.batch)
 
 
-class RocksDBIterator(object):
+class RocksDBIterator:
     '''An iterator for RocksDB.'''
 
     def __init__(self, db, prefix, reverse):
