@@ -1,3 +1,4 @@
+import dataclasses
 import datetime
 import logging
 import os
@@ -6,13 +7,13 @@ from functools import partial
 from random import randrange, choice, seed
 
 import pytest
-from aiorpcx import Event, TaskGroup, sleep, spawn, ignore_after
+from aiorpcx import Event, sleep, ignore_after
 
 from electrumx.server.mempool import MemPool, MemPoolAPI
 from electrumx.lib.coins import BitcoinCash
 from electrumx.lib.hash import HASHX_LEN, hex_str_to_hash, hash_to_hex_str
 from electrumx.lib.tx import Tx, TxInput, TxOutput
-from electrumx.lib.util import make_logger
+from electrumx.lib.util import OldTaskGroup
 
 
 coin = BitcoinCash
@@ -267,7 +268,7 @@ async def test_keep_synchronized(caplog):
     mempool = MemPool(coin, api)
     event = Event()
     with caplog.at_level(logging.INFO):
-        async with TaskGroup() as group:
+        async with OldTaskGroup() as group:
             await group.spawn(mempool.keep_synchronized, event)
             await event.wait()
             await group.cancel_remaining()
@@ -286,7 +287,7 @@ async def test_balance_delta():
     api.initialize()
     mempool = MemPool(coin, api)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         await group.cancel_remaining()
@@ -309,7 +310,7 @@ async def test_compact_fee_histogram():
     api.initialize()
     mempool = MemPool(coin, api)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         await group.cancel_remaining()
@@ -324,13 +325,44 @@ async def test_compact_fee_histogram():
     assert all(rates[n] < rates[n - 1] for n in range(1, len(rates)))
 
 
+def test_compress_histogram():
+    histogram = {
+        10: 100_000,
+        11: 1_000,
+        12: 10_000_000,
+        13: 1_000,
+        14: 1_000,
+        15: 1_000,
+        16: 1_000,
+        17: 1_000,
+        18: 900_000,
+        19: 1_000,
+        20: 1_000,
+        21: 75_000,
+        22: 1_000,
+    }
+    compact = MemPool._compress_histogram(histogram, bin_size=100_000)
+    assert compact == [(19, 78000), (18, 900000), (13, 5000), (12, 10000000)]
+
+    histogram = {
+        1.0: 10_000_000,
+        1.1: 30_000,
+        1.2: 40_000,
+        10: 500_000,
+        10.1: 1_000,
+        11: 50_000,
+    }
+    compact = MemPool._compress_histogram(histogram, bin_size=100_000)
+    assert compact == [(10.1, 51000), (10, 500000), (1.1, 70000), (1.0, 10000000)]
+
+
 @pytest.mark.asyncio
 async def test_potential_spends():
     api = API()
     api.initialize()
     mempool = MemPool(coin, api)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         await group.cancel_remaining()
@@ -365,7 +397,7 @@ async def test_transaction_summaries(caplog):
     mempool = MemPool(coin, api)
     event = Event()
     with caplog.at_level(logging.INFO):
-        async with TaskGroup() as group:
+        async with OldTaskGroup() as group:
             await group.spawn(mempool.keep_synchronized, event)
             await event.wait()
             await group.cancel_remaining()
@@ -385,7 +417,7 @@ async def test_unordered_UTXOs():
     api.initialize()
     mempool = MemPool(coin, api)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         await group.cancel_remaining()
@@ -400,7 +432,10 @@ async def test_unordered_UTXOs():
     for hashX in api.hashXs:
         mempool_result = await mempool.unordered_UTXOs(hashX)
         our_result = utxos.get(hashX, [])
-        assert set(our_result) == set(mempool_result)
+        assert set(our_result) == {
+            dataclasses.astuple(mr)
+            for mr in mempool_result
+        }
 
 
 @pytest.mark.asyncio
@@ -409,7 +444,7 @@ async def test_mempool_removals():
     api.initialize()
     mempool = MemPool(coin, api, refresh_secs=0.01)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         # Remove half the TXs from the mempool
@@ -439,7 +474,7 @@ async def test_daemon_drops_txs():
     api.initialize()
     mempool = MemPool(coin, api, refresh_secs=0.01)
     event = Event()
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         await group.spawn(mempool.keep_synchronized, event)
         await event.wait()
         await _test_summaries(mempool, api)
@@ -467,7 +502,7 @@ async def test_notifications(caplog):
 
     caplog.set_level(logging.DEBUG)
 
-    async with TaskGroup() as group:
+    async with OldTaskGroup() as group:
         # First batch enters the mempool
         api.raw_txs = {hash: raw_txs[hash] for hash in first_hashes}
         api.txs = {hash: txs[hash] for hash in first_hashes}
@@ -495,7 +530,7 @@ async def test_notifications(caplog):
         api.txs = {hash: txs[hash] for hash in second_hashes}
         # Delay the DB update
         assert not in_caplog(caplog, 'waiting for DB to sync')
-        async with ignore_after(mempool.refresh_secs * 2):
+        async with ignore_after(max(mempool.refresh_secs * 2, 0.5)):
             await event.wait()
         assert in_caplog(caplog, 'waiting for DB to sync')
         assert len(api.on_mempool_calls) == 2
@@ -527,7 +562,7 @@ async def test_dropped_txs(caplog):
             del api.txs[prev_hash]
 
     with caplog.at_level(logging.INFO):
-        async with TaskGroup() as group:
+        async with OldTaskGroup() as group:
             await group.spawn(mempool.keep_synchronized, event)
             await event.wait()
             await group.cancel_remaining()
